+++ gusty/importing.py
@@ -1,0 +1,61 @@
+import os, sys, pkgutil, itertools, airflow
+from inflection import underscore
+
+############
+## Params ##
+############
+
+airflow_version = int(str(airflow.__version__)[0])
+valid_extensions = (".yml", ".Rmd", ".ipynb")
+
+###########################
+## Operator Import Logic ##
+###########################
+
+
+def get_operator_location(operator_string):
+    # location will generally be 'airflow',
+    # but if it's 'local', then we will look locally for the operator
+    return operator_string.split(".")[0]
+
+
+def get_operator_name(operator_string):
+    # the actual name of the operator
+    return operator_string.split(".")[-1]
+
+
+def get_operator_module(operator_string):
+    # the module, for when the operator is not a local operator
+    operator_path = ".".join(operator_string.split(".")[:-1])
+    assert len(operator_path) != 0, (
+        "Please specify a format like 'package.operator' to specify your operator. You passed in '%s'"
+        % operator_string
+    )
+    return operator_path
+
+
+# Add $AIRFLOW_HOME/operators directory to path for local.operator syntax to work
+CUSTOM_OPERATORS_DIR = os.path.join(os.environ.get("AIRFLOW_HOME", ""), "operators")
+sys.path.append(CUSTOM_OPERATORS_DIR)
+module_paths = [("", [CUSTOM_OPERATORS_DIR])]
+pairs = [
+    [(_.name, m + _.name) for _ in pkgutil.iter_modules(path)]
+    for m, path in module_paths
+]
+module_dict = dict(itertools.chain(*pairs))
+
+
+def get_operator(operator_string):
+    operator_name = get_operator_name(operator_string)
+    operator_location = get_operator_location(operator_string)
+
+    module_name = (
+        module_dict[underscore(operator_name)]
+        if operator_location == "local"
+        else get_operator_module(operator_string)
+    )
+
+    import_stmt = "from %s import %s" % (module_name, operator_name)
+
+    exec(import_stmt)
+    return eval(operator_name)
